+++ doubleclick-openrtb/src/main/java/com/google/doubleclick/openrtb/DoubleClickOpenRtbMapper.java
@@ -77,7 +77,6 @@ import java.util.EnumSet;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.regex.Pattern;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
@@ -95,9 +94,6 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
     NetworkBid.BidRequest, NetworkBid.BidResponse,
     NetworkBid.BidRequest.Builder, NetworkBid.BidResponse.Builder> {
   private static final Logger logger = LoggerFactory.getLogger(DoubleClickOpenRtbMapper.class);
-  private static final String YOUTUBE_AFV_USER_ID = "afv_user_id_";
-  private static final Pattern SEMITRANSPARENT_CHANNEL =
-      Pattern.compile("pack-(brand|semi|anon)-([^\\-]+)::(.+)");
   private static final Joiner csvJoiner = Joiner.on(",").skipNulls();
   private static final int MICROS_PER_CURRENCY_UNIT = 1_000_000;
 
@@ -435,23 +431,21 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
       mapped = true;
     }
 
-    String channelId = findChannelId(dcRequest);
-    if (channelId != null) {
-      app.setId(channelId);
-      mapped = true;
-    }
-
     Publisher.Builder pub = buildPublisher(dcRequest);
     if (pub != null) {
       app.setPublisher(pub);
       mapped = true;
     }
 
-    for (ExtMapper extMapper : extMappers) {
-      mapped |= extMapper.toOpenRtbApp(dcRequest, app);
+    if (!mapped) {
+      return null;
     }
 
-    return mapped ? app : null;
+    for (ExtMapper extMapper : extMappers) {
+      extMapper.toOpenRtbApp(dcRequest, app);
+    }
+
+    return app;
   }
 
   @Nullable protected Content.Builder buildAppContent(NetworkBid.BidRequest dcRequest) {
@@ -507,17 +501,15 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
       mapped = true;
     }
 
-    String channelId = findChannelId(dcRequest);
-    if (channelId != null) {
-      site.setId(channelId);
-      mapped = true;
+    if (!mapped) {
+      return null;
     }
 
     for (ExtMapper extMapper : extMappers) {
-      mapped |= extMapper.toOpenRtbSite(dcRequest, site);
+      extMapper.toOpenRtbSite(dcRequest, site);
     }
 
-    return mapped ? site : null;
+    return site;
   }
 
   @Nullable protected Content.Builder buildContent(NetworkBid.BidRequest dcRequest) {
@@ -887,20 +879,6 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
     }
 
     return user;
-  }
-
-  @Nullable protected String findChannelId(NetworkBid.BidRequest dcRequest) {
-    for (NetworkBid.BidRequest.AdSlot dcSlot : dcRequest.getAdslotList()) {
-      for (String dcChannel : dcSlot.getTargetableChannelList()) {
-        if (dcChannel.startsWith(YOUTUBE_AFV_USER_ID)) {
-          return dcChannel.substring(YOUTUBE_AFV_USER_ID.length());
-        } else if (SEMITRANSPARENT_CHANNEL.matcher(dcChannel).matches()) {
-          return dcChannel;
-        }
-      }
-    }
-
-    return null;
   }
 
   @Override public NetworkBid.BidResponse.Builder toExchangeBidResponse(